Improve `truss watch`: performance improvements and more logging (#808)
With this PR, we no longer spawn a background thread for detecting changes and creating patches. This should 
mitigate the issue where another thread unnecessarily consumes CPU cycles and slows down truss watch.

In addition to the speedups, removing TrussFileSyncer involved changing the logger used in BasetenRemote.patch. This 
fixes a bug that resulted in the info logs in BasetenRemote.patch not being logged.