Improve `truss watch`: respect `.truss_ignore` when calculating patches (#812)
Although we currently account for truss ignore patterns when detecting file changes with watchfiles.watch, we do not 
account for the ignore patterns when calculating patches or the truss hash and signature. This means that changes in 
ignored files will still be patched. This PR fixes that, along with some refactoring and bug fixes.